+++ actionista/__init__.py
@@ -1,3 +1,3 @@
 
 
-__version__ = '2019.09.06'  # remember to also update setup.py
+__version__ = '2019.09.09'  # remember to also update setup.py